+++ History.md
@@ -1,3 +1,8 @@
+0.2.0 - August 20, 2013
+-----------------------
+* move `template` to the prototype
+* add other class helpers via [`ianstormtaylor/classes`](https://github.com/ianstormtaylor/classes)
+
 0.1.0 - August 3, 2013
 ----------------------
 * change `#view` to `#View`